requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.6
pydantic==2.4.2
gnews==0.2.1
transformers==4.36.2
torch==2.1.2
sentencepiece==0.1.99 